Refactor SpectreConsoleDemo into focused files
Split the monolithic Program.cs into three concerns:

* Program.cs - bare entry point. Parses --auto/--headless via System.CommandLine 2.0.2 (matching Hex1b.Tool / Hex1b.PtyHost), wires up the terminal builder, and orchestrates the optional automator. ~85 lines.
* InteractiveDemo.cs - the self-contained Spectre.Console showcase: menu loop, the ten demo handlers (markup, tables, trees, calendar, charts, status, live, progress, prompts), and WaitForEnterAsync. Has no Hex1b dependency at all - it only sees IAnsiConsole.
* DemoAutomator.cs - the Hex1bTerminalAutomator script that drives the demo end-to-end, including DrivePromptsAsync.

MenuItems is exposed as a public static IReadOnlyList<string> on InteractiveDemo so the automator references the canonical list rather than duplicating it.

Switching to System.CommandLine also gives us auto-generated --help / --version output.
